+++ lib/vizcore/dsl.rb
@@ -8,6 +8,7 @@ end
 
 require_relative "dsl/file_watcher"
 require_relative "dsl/mapping_resolver"
+require_relative "dsl/mapping_transform_builder"
 require_relative "dsl/midi_map_executor"
 require_relative "dsl/reaction_builder"
 require_relative "dsl/layer_builder"